autodl-fs/busi_ssl_baseline/busi_ssl_baseline/requirements.txt
# Python 3.10.14
torch==2.4.0
torchvision==0.19.0
segmentation-models-pytorch==0.5.0
PyYAML==6.0.2
numpy==1.26.4
pandas==2.2.2
Pillow==10.4.0
tqdm==4.66.5
tensorboard==2.17.1
